working version: - need genesis with right txs - fixed validators addresses - add proposals and votes - add delegation and staking
