+++ cmd/gaia/lcd_test/lcd_test.go
@@ -1,4 +1,4 @@
-package lcd
+package lcdtest
 
 import (
 	"encoding/base64"
@@ -53,7 +53,7 @@ func TestVersion(t *testing.T) {
 		t.SkipNow()
 	}
 
-	cleanup, _, _, port := InitializeTestLCD(t, 1, []sdk.AccAddress{}, true)
+	cleanup, _, _, port := InitializeLCD(1, []sdk.AccAddress{}, true)
 	defer cleanup()
 
 	// node info
@@ -76,14 +76,14 @@ func TestVersion(t *testing.T) {
 }
 
 func TestNodeStatus(t *testing.T) {
-	cleanup, _, _, port := InitializeTestLCD(t, 1, []sdk.AccAddress{}, true)
+	cleanup, _, _, port := InitializeLCD(1, []sdk.AccAddress{}, true)
 	defer cleanup()
 	getNodeInfo(t, port)
 	getSyncStatus(t, port, false)
 }
 
 func TestBlock(t *testing.T) {
-	cleanup, _, _, port := InitializeTestLCD(t, 1, []sdk.AccAddress{}, true)
+	cleanup, _, _, port := InitializeLCD(1, []sdk.AccAddress{}, true)
 	defer cleanup()
 	getBlock(t, port, -1, false)
 	getBlock(t, port, 2, false)
@@ -91,7 +91,7 @@ func TestBlock(t *testing.T) {
 }
 
 func TestValidators(t *testing.T) {
-	cleanup, _, _, port := InitializeTestLCD(t, 1, []sdk.AccAddress{}, true)
+	cleanup, _, _, port := InitializeLCD(1, []sdk.AccAddress{}, true)
 	defer cleanup()
 	resultVals := getValidatorSets(t, port, -1, false)
 	require.Contains(t, resultVals.Validators[0].Address.String(), "cosmosvalcons")
@@ -105,7 +105,7 @@ func TestCoinSend(t *testing.T) {
 	require.NoError(t, err)
 	addr, seed, err := CreateAddr(name1, pw, kb)
 	require.NoError(t, err)
-	cleanup, _, _, port := InitializeTestLCD(t, 1, []sdk.AccAddress{addr}, true)
+	cleanup, _, _, port := InitializeLCD(1, []sdk.AccAddress{addr}, true)
 	defer cleanup()
 
 	bz, err := hex.DecodeString("8FA6AB57AD6870F6B5B2E57735F38F2F30E73CB6")
@@ -197,7 +197,7 @@ func TestCoinSendAccAuto(t *testing.T) {
 	require.NoError(t, err)
 	addr, seed, err := CreateAddr(name1, pw, kb)
 	require.NoError(t, err)
-	cleanup, _, _, port := InitializeTestLCD(t, 1, []sdk.AccAddress{addr}, true)
+	cleanup, _, _, port := InitializeLCD(1, []sdk.AccAddress{addr}, true)
 	defer cleanup()
 
 	acc := getAccount(t, port, addr)
@@ -223,7 +223,7 @@ func TestCoinMultiSendGenerateOnly(t *testing.T) {
 	require.NoError(t, err)
 	addr, seed, err := CreateAddr(name1, pw, kb)
 	require.NoError(t, err)
-	cleanup, _, _, port := InitializeTestLCD(t, 1, []sdk.AccAddress{addr}, true)
+	cleanup, _, _, port := InitializeLCD(1, []sdk.AccAddress{addr}, true)
 	defer cleanup()
 
 	// generate only
@@ -247,7 +247,7 @@ func TestCoinSendGenerateSignAndBroadcast(t *testing.T) {
 	require.NoError(t, err)
 	addr, seed, err := CreateAddr(name1, pw, kb)
 	require.NoError(t, err)
-	cleanup, _, _, port := InitializeTestLCD(t, 1, []sdk.AccAddress{addr}, true)
+	cleanup, _, _, port := InitializeLCD(1, []sdk.AccAddress{addr}, true)
 
 	defer cleanup()
 	acc := getAccount(t, port, addr)
@@ -291,7 +291,7 @@ func TestEncodeTx(t *testing.T) {
 	require.NoError(t, err)
 	addr, seed, err := CreateAddr(name1, pw, kb)
 	require.NoError(t, err)
-	cleanup, _, _, port := InitializeTestLCD(t, 1, []sdk.AccAddress{addr}, true)
+	cleanup, _, _, port := InitializeLCD(1, []sdk.AccAddress{addr}, true)
 	defer cleanup()
 
 	res, body, _ := doTransferWithGas(t, port, seed, name1, memo, "", addr, "2", 1, false, false, fees)
@@ -326,7 +326,7 @@ func TestTxs(t *testing.T) {
 	require.NoError(t, err)
 	addr, seed, err := CreateAddr(name1, pw, kb)
 	require.NoError(t, err)
-	cleanup, _, _, port := InitializeTestLCD(t, 1, []sdk.AccAddress{addr}, true)
+	cleanup, _, _, port := InitializeLCD(1, []sdk.AccAddress{addr}, true)
 	defer cleanup()
 
 	var emptyTxs []sdk.TxResponse
@@ -379,7 +379,7 @@ func TestPoolParamsQuery(t *testing.T) {
 	require.NoError(t, err)
 	addr, _, err := CreateAddr(name1, pw, kb)
 	require.NoError(t, err)
-	cleanup, _, _, port := InitializeTestLCD(t, 1, []sdk.AccAddress{addr}, true)
+	cleanup, _, _, port := InitializeLCD(1, []sdk.AccAddress{addr}, true)
 	defer cleanup()
 
 	defaultParams := staking.DefaultParams()
@@ -404,7 +404,7 @@ func TestPoolParamsQuery(t *testing.T) {
 }
 
 func TestValidatorsQuery(t *testing.T) {
-	cleanup, valPubKeys, operAddrs, port := InitializeTestLCD(t, 1, []sdk.AccAddress{}, true)
+	cleanup, valPubKeys, operAddrs, port := InitializeLCD(1, []sdk.AccAddress{}, true)
 	defer cleanup()
 
 	require.Equal(t, 1, len(valPubKeys))
@@ -424,7 +424,7 @@ func TestValidatorsQuery(t *testing.T) {
 }
 
 func TestValidatorQuery(t *testing.T) {
-	cleanup, valPubKeys, operAddrs, port := InitializeTestLCD(t, 1, []sdk.AccAddress{}, true)
+	cleanup, valPubKeys, operAddrs, port := InitializeLCD(1, []sdk.AccAddress{}, true)
 	defer cleanup()
 	require.Equal(t, 1, len(valPubKeys))
 	require.Equal(t, 1, len(operAddrs))
@@ -439,7 +439,7 @@ func TestBonding(t *testing.T) {
 	addr, _, err := CreateAddr(name1, pw, kb)
 	require.NoError(t, err)
 
-	cleanup, valPubKeys, operAddrs, port := InitializeTestLCD(t, 2, []sdk.AccAddress{addr}, false)
+	cleanup, valPubKeys, operAddrs, port := InitializeLCD(2, []sdk.AccAddress{addr}, false)
 	tests.WaitForHeight(1, port)
 	defer cleanup()
 
@@ -603,7 +603,7 @@ func TestSubmitProposal(t *testing.T) {
 	require.NoError(t, err)
 	addr, seed, err := CreateAddr(name1, pw, kb)
 	require.NoError(t, err)
-	cleanup, _, _, port := InitializeTestLCD(t, 1, []sdk.AccAddress{addr}, true)
+	cleanup, _, _, port := InitializeLCD(1, []sdk.AccAddress{addr}, true)
 	defer cleanup()
 
 	acc := getAccount(t, port, addr)
@@ -641,7 +641,7 @@ func TestSubmitParamChangeProposal(t *testing.T) {
 	require.NoError(t, err)
 	addr, seed, err := CreateAddr(name1, pw, kb)
 	require.NoError(t, err)
-	cleanup, _, _, port := InitializeTestLCD(t, 1, []sdk.AccAddress{addr}, true)
+	cleanup, _, _, port := InitializeLCD(1, []sdk.AccAddress{addr}, true)
 	defer cleanup()
 
 	acc := getAccount(t, port, addr)
@@ -679,7 +679,7 @@ func TestDeposit(t *testing.T) {
 	require.NoError(t, err)
 	addr, seed, err := CreateAddr(name1, pw, kb)
 	require.NoError(t, err)
-	cleanup, _, _, port := InitializeTestLCD(t, 1, []sdk.AccAddress{addr}, true)
+	cleanup, _, _, port := InitializeLCD(1, []sdk.AccAddress{addr}, true)
 	defer cleanup()
 
 	acc := getAccount(t, port, addr)
@@ -739,7 +739,7 @@ func TestVote(t *testing.T) {
 	require.NoError(t, err)
 	addr, seed, err := CreateAddr(name1, pw, kb)
 	require.NoError(t, err)
-	cleanup, _, operAddrs, port := InitializeTestLCD(t, 1, []sdk.AccAddress{addr}, true)
+	cleanup, _, operAddrs, port := InitializeLCD(1, []sdk.AccAddress{addr}, true)
 	defer cleanup()
 
 	acc := getAccount(t, port, addr)
@@ -827,7 +827,7 @@ func TestUnjail(t *testing.T) {
 	require.NoError(t, err)
 	addr, _, err := CreateAddr(name1, pw, kb)
 	require.NoError(t, err)
-	cleanup, valPubKeys, _, port := InitializeTestLCD(t, 1, []sdk.AccAddress{addr}, true)
+	cleanup, valPubKeys, _, port := InitializeLCD(1, []sdk.AccAddress{addr}, true)
 	defer cleanup()
 
 	// XXX: any less than this and it fails
@@ -847,7 +847,7 @@ func TestProposalsQuery(t *testing.T) {
 	require.NoError(t, err)
 	addrs, seeds, names, passwords := CreateAddrs(kb, 2)
 
-	cleanup, _, _, port := InitializeTestLCD(t, 1, []sdk.AccAddress{addrs[0], addrs[1]}, true)
+	cleanup, _, _, port := InitializeLCD(1, []sdk.AccAddress{addrs[0], addrs[1]}, true)
 	defer cleanup()
 
 	depositParam := getDepositParam(t, port)
@@ -973,7 +973,7 @@ func TestProposalsQuery(t *testing.T) {
 }
 
 func TestSlashingGetParams(t *testing.T) {
-	cleanup, _, _, port := InitializeTestLCD(t, 1, []sdk.AccAddress{}, true)
+	cleanup, _, _, port := InitializeLCD(1, []sdk.AccAddress{}, true)
 	defer cleanup()
 
 	res, body := Request(t, port, "GET", "/slashing/parameters", nil)
@@ -985,7 +985,7 @@ func TestSlashingGetParams(t *testing.T) {
 }
 
 func TestDistributionGetParams(t *testing.T) {
-	cleanup, _, _, port := InitializeTestLCD(t, 1, []sdk.AccAddress{}, true)
+	cleanup, _, _, port := InitializeLCD(1, []sdk.AccAddress{}, true)
 	defer cleanup()
 
 	res, body := Request(t, port, "GET", "/distribution/parameters", nil)
@@ -998,7 +998,7 @@ func TestDistributionFlow(t *testing.T) {
 	require.NoError(t, err)
 	addr, seed, err := CreateAddr(name1, pw, kb)
 	require.NoError(t, err)
-	cleanup, _, valAddrs, port := InitializeTestLCD(t, 1, []sdk.AccAddress{addr}, true)
+	cleanup, _, valAddrs, port := InitializeLCD(1, []sdk.AccAddress{addr}, true)
 	defer cleanup()
 
 	valAddr := valAddrs[0]
@@ -1075,7 +1075,7 @@ func TestMintingQueries(t *testing.T) {
 	require.NoError(t, err)
 	addr, _, err := CreateAddr(name1, pw, kb)
 	require.NoError(t, err)
-	cleanup, _, _, port := InitializeTestLCD(t, 1, []sdk.AccAddress{addr}, true)
+	cleanup, _, _, port := InitializeLCD(1, []sdk.AccAddress{addr}, true)
 	defer cleanup()
 
 	res, body := Request(t, port, "GET", "/minting/parameters", nil)
@@ -1102,7 +1102,7 @@ func TestAccountBalanceQuery(t *testing.T) {
 	require.NoError(t, err)
 	addr, _, err := CreateAddr(name1, pw, kb)
 	require.NoError(t, err)
-	cleanup, _, _, port := InitializeTestLCD(t, 1, []sdk.AccAddress{addr}, true)
+	cleanup, _, _, port := InitializeLCD(1, []sdk.AccAddress{addr}, true)
 	defer cleanup()
 
 	bz, err := hex.DecodeString("8FA6AB57AD6870F6B5B2E57735F38F2F30E73CB6")
